Move package from "example" company domain to "vmware" company domain
