--- app/src/main/java/com/example/nimbus/data/LoginDataSource.java
@@ -1,29 +1,0 @@
-package com.example.nimbus.data;
-
-import com.example.nimbus.data.model.LoggedInUser;
-
-import java.io.IOException;
-
-/**
- * Class that handles authentication w/ login credentials and retrieves user information.
- */
-public class LoginDataSource {
-
-    public Result<LoggedInUser> login(String password) {
-
-        try {
-            // TODO: handle loggedInUser authentication
-            LoggedInUser fakeUser =
-                    new LoggedInUser(
-                            java.util.UUID.randomUUID().toString(),
-                            "Jane Doe");
-            return new Result.Success<>(fakeUser);
-        } catch (Exception e) {
-            return new Result.Error(new IOException("Error logging in", e));
-        }
-    }
-
-    public void logout() {
-        // TODO: revoke authentication
-    }
-}
