--- app/src/main/java/com/example/nimbus/data/model/BlueprintsModel.java
@@ -1,13 +1,0 @@
-package com.example.nimbus.data.model;
-
-public class BlueprintsModel {
-    public String helloData;
-    public String worldData;
-    public String dataIndex;
-
-    public BlueprintsModel(String helloData, String worldData, String dataIndex) {
-        this.helloData = helloData;
-        this.worldData = worldData;
-        this.dataIndex = dataIndex;
-    }
-}
