--- app/src/main/java/com/example/nimbus/data/model/DeploymentsModel.java
@@ -1,13 +1,0 @@
-package com.example.nimbus.data.model;
-
-public class DeploymentsModel {
-    public String helloData;
-    public String worldData;
-    public String dataIndex;
-
-    public DeploymentsModel(String helloData, String worldData, String dataIndex) {
-        this.helloData = helloData;
-        this.worldData = worldData;
-        this.dataIndex = dataIndex;
-    }
-}
