--- app/src/main/java/com/example/nimbus/ui/main/BlueprintsViewFragment.java
@@ -1,84 +1,0 @@
-package com.example.nimbus.ui.main;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProviders;
-import androidx.recyclerview.widget.DefaultItemAnimator;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.example.nimbus.R;
-import com.example.nimbus.data.model.BlueprintsModel;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * A placeholder fragment containing a simple view.
- */
-public class BlueprintsViewFragment extends Fragment {
-
-    private static final String ARG_SECTION_NUMBER = "section_number";
-
-    private PageViewModel pageViewModel;
-    private BlueprintsViewModel mViewModel;
-
-    private RecyclerView recyclerView;
-
-    private BlueprintsAdapter rvAdapter;
-    private List<BlueprintsModel> blueprintsTests;
-
-
-    public static BlueprintsViewFragment newInstance(int index) {
-        BlueprintsViewFragment fragment = new BlueprintsViewFragment();
-        Bundle bundle = new Bundle();
-        bundle.putInt(ARG_SECTION_NUMBER, index);
-        fragment.setArguments(bundle);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        mViewModel = ViewModelProviders.of(this).get(BlueprintsViewModel.class);
-        pageViewModel = ViewModelProviders.of(this).get(PageViewModel.class);
-        int index = 1;
-        if (getArguments() != null) {
-            index = getArguments().getInt(ARG_SECTION_NUMBER);
-        }
-        pageViewModel.setIndex(index);
-
-    }
-
-    @Override
-    public View onCreateView(
-            @NonNull LayoutInflater inflater, ViewGroup container,
-            Bundle savedInstanceState) {
-        View root = inflater.inflate(R.layout.fragment_blueprints, container, false);
-        final TextView textView = root.findViewById(R.id.section_label);
-
-        recyclerView = root.findViewById(R.id.fragment_blueprints_recycler);
-        LinearLayoutManager llm = new LinearLayoutManager(getActivity(), LinearLayoutManager.VERTICAL, false);
-        recyclerView.setLayoutManager(llm);
-        blueprintsTests = new ArrayList<>();
-        blueprintsTests = mViewModel.initializeBlueprintsData();
-        rvAdapter = new BlueprintsAdapter(blueprintsTests);
-
-        recyclerView.setAdapter(rvAdapter);
-        recyclerView.setItemAnimator(new DefaultItemAnimator());
-//        pageViewModel.getText().observe(this, new Observer<String>() {
-//            @Override
-//            public void onChanged(@Nullable String s) {
-//                textView.setText(s);
-//            }
-//        });
-        return recyclerView;
-    }
-
-}
