+++ app/src/main/java/com/vmware/nimbus/data/model/DeploymentsModel.java
@@ -1,0 +1,13 @@
+package com.vmware.nimbus.data.model;
+
+public class DeploymentsModel {
+    public String helloData;
+    public String worldData;
+    public String dataIndex;
+
+    public DeploymentsModel(String helloData, String worldData, String dataIndex) {
+        this.helloData = helloData;
+        this.worldData = worldData;
+        this.dataIndex = dataIndex;
+    }
+}
