+++ app/src/main/java/com/vmware/nimbus/ui/login/LoggedInUserView.java
@@ -1,0 +1,17 @@
+package  com.vmware.nimbus.ui.login;
+
+/**
+ * Class exposing authenticated user details to the UI.
+ */
+class LoggedInUserView {
+    private String displayName;
+    //... other data fields that may be accessible to the UI
+
+    LoggedInUserView(String displayName) {
+        this.displayName = displayName;
+    }
+
+    String getDisplayName() {
+        return displayName;
+    }
+}
